Updated Resturant page
SO the menu has been updated specifially the
appetizers portion by adding more food, wrapping
so the page does not overflow the margins of the
page and repeat of the background image so the
image repeats as the contents begins to grow.
what to do know is finish the menu variety
by adding desserts and entrees and then proceeding
to finsih the last two button which are about and
services and book a table which wont do much and
then updating the branhes and applying the infor
mation form the classes branch into the main branch
this has been a fun project and im missing the
tic tac toe project lowkey. <= slang